crypto_ledger/requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==23.1.0
awsebcli==3.20.7
Babel==2.12.1
bcrypt==4.0.1
billiard==3.6.4.0
blessed==1.20.0
boto3==1.26.135
botocore==1.29.135
cached-property==1.5.2
celery==5.2.7
cement==2.8.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.1
django-allauth==0.54.0
django-appconf==1.0.5
django-cors-headers==4.0.0
django-filter==22.1
django-money==3.1.0
django-rest-registration==0.8.1
django-storages==1.13.2
django-unitology==0.0.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.1.0
idna==3.4
jmespath==1.0.1
jsonschema==3.2.0
kombu==5.2.4
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.3
paramiko==3.3.1
pathspec==0.10.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.7
py-moneyed==2.0
pycparser==2.21
PyJWT==2.7.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==5.4.1
quantities==0.14.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.4
stripe==5.5.0
termcolor==1.1.0
texttable==1.6.7
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.1.9
websocket-client==0.59.0
django-health-check==3.17.0
whitenoise==6.5.0

